reid_embedding_extractor/requirements.txt
# ReID Embedding Extractor 의존성 패키지 - PyTorch 최신 버전 호환

# PyTorch 및 관련 패키지 (최신 버전)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 이미지 처리
Pillow>=9.0.0
opencv-python>=4.7.0

# 수치 계산
numpy>=1.21.0
scipy>=1.9.0

# 설정 관리
yacs>=0.1.8

# 유틸리티
tqdm>=4.64.0

# 로깅 및 기타
colorlog>=6.7.0
timm>=0.9.0

# 선택적 패키지 (성능 향상)
# faiss-cpu>=1.7.4  # CPU 버전
# faiss-gpu>=1.7.4  # GPU 버전 (CUDA 지원 시)

# 개발 및 테스트용
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0